refactor: eliminate _extended test files by renaming and merging
Renamed 9 standalone extended files (git mv) and merged 4 extended files
into their base counterparts, deduplicating helpers and consolidating all
test classes under clean names with no _extended suffix.
