+++ .secrets.baseline
@@ -486,14 +486,14 @@
         "filename": "tests/unit/test_hive_auth_async.py",
         "hashed_secret": "5c5a15a8b0b3e154d77746945e563ba40100681b",
         "is_verified": false,
-        "line_number": 165
+        "line_number": 173
       },
       {
         "type": "Secret Keyword",
         "filename": "tests/unit/test_hive_auth_async.py",
         "hashed_secret": "d8bce9746547bb7743e5933fbf0fc4f2d2cbcad3",
         "is_verified": false,
-        "line_number": 221
+        "line_number": 229
       }
     ],
     "tests/unit/test_hive_auth_async_extended.py": [
@@ -580,5 +580,5 @@
       }
     ]
   },
-  "generated_at": "2026-05-25T16:53:38Z"
+  "generated_at": "2026-05-27T18:17:19Z"
 }
