+++ tests/unit/test_sensor.py
