+++ tests/unit/test_session_auth.py
